DMM-019 moved onclick into fragment to fix bug, where button wasn´t clickable after fragment change [II, MM]
