+++ app/src/main/res/layout/main_activity.xml
@@ -35,4 +35,4 @@
             app:navGraph="@navigation/mobile_navigation" />
 
 
-    </androidx.constraintlayout.widget.ConstraintLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
